rename packages and add homework from lecture 3 equal lecture 5
