+++ src/test/java/ASE/lecture3/ex/AseTestLecture3.java
@@ -1,4 +1,4 @@
-package ASE.lekce3.cv;
+package ASE.lecture3.ex;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
@@ -7,7 +7,7 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.annotations.Test;
 
-public class AseTestLekce3 {
+public class AseTestLecture3 {
     private WebDriverWait wait;
