--- src/test/java/ASE/lekce3/du/AseTestLekce3_DU_Xpath.java
@@ -1,22 +1,0 @@
-//package ASE.lekce3.du;
-//
-//import org.openqa.selenium.WebDriver;
-//import org.openqa.selenium.chrome.ChromeDriver;
-//import org.testng.annotations.AfterTest;
-//import org.testng.annotations.Test;
-//
-//public class AseTestLekce3_DU_Xpath {
-//
-//    @AfterTest
-//    driver.quit();
-//
-//    @Test
-//    public void registerToTBD(){
-//        WebDriver driver = new ChromeDriver() {
-//
-//        }
-//
-//
-//    }
-//
-//}
